+++ src/main/java/com/verdantartifice/primalmagic/common/events/EntityEvents.java
@@ -9,8 +9,12 @@ import com.verdantartifice.primalmagic.common.stats.StatsManager;
 import com.verdantartifice.primalmagic.common.stats.StatsPM;
 
 import net.minecraft.util.Mth;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.damagesource.EntityDamageSource;
+import net.minecraft.world.damagesource.IndirectEntityDamageSource;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.animal.Wolf;
 import net.minecraft.world.entity.player.Player;
@@ -22,6 +26,7 @@ import net.minecraftforge.event.entity.EntityTeleportEvent;
 import net.minecraftforge.event.entity.living.AnimalTameEvent;
 import net.minecraftforge.event.entity.living.BabyEntitySpawnEvent;
 import net.minecraftforge.event.entity.living.LivingEntityUseItemEvent;
+import net.minecraftforge.event.entity.living.LootingLevelEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -114,6 +119,23 @@ public class EntityEvents {
             MobEffectInstance effectInstance = entity.getEffect(MobEffects.DAMAGE_RESISTANCE);
             int amplifier = (effectInstance == null) ? 0 : Mth.clamp(1 + effectInstance.getAmplifier(), 0, enchantLevel - 1);
             entity.addEffect(new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, 10, amplifier));
+        }
+    }
+    
+    @SubscribeEvent
+    public static void onLootingLevel(LootingLevelEvent event) {
+        // If the damage was magical, apply the Treasure enchantment as a looting modifier if greater than what's already there
+        DamageSource source = event.getDamageSource();
+        Entity caster = null;
+        if (source.isMagic()) {
+            if (source instanceof IndirectEntityDamageSource indirectSource) {
+                caster = indirectSource.getEntity();
+            } else if (source instanceof EntityDamageSource directSource) {
+                caster = directSource.getEntity();
+            }
+        }
+        if (caster != null && caster instanceof LivingEntity living) {
+            event.setLootingLevel(Math.max(event.getLootingLevel(), EnchantmentHelper.getEnchantmentLevel(EnchantmentsPM.TREASURE.get(), living)));
         }
     }
 }